requirements.txt
agno
annotated-types
anyio
beautifulsoup4
certifi
charset-normalizer
click
colorama
distro
dnspython
docstring_parser
duckduckgo_search
email_validator
fastapi
fastapi-cli
frozendict
fuzzywuzzy
gitdb
GitPython
greenlet
h11
httpcore
httptools
httpx
idna
Jinja2
jiter
Levenshtein
lxml
markdown-it-py
MarkupSafe
mdurl
multitasking
numpy
openai
pandas
peewee
platformdirs
primp
psycopg2
pydantic
pydantic-settings
pydantic_core
Pygments
python-dateutil
python-dotenv
python-Levenshtein
python-multipart
pytz
PyYAML
RapidFuzz
requests
rich
rich-toolkit
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
starlette
tomli
tqdm
typer
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
watchfiles
websockets
yfinance
python-dotenv